+++ lib/pub_subx/auto.ex
@@ -36,7 +36,7 @@ defmodule PubSubx.Auto do
       def child_spec(opts) do
         %{
           id: __MODULE__,
-          start: {__MODULE__, :start_link, opts}
+          start: {__MODULE__, :start_link, [opts]}
         }
       end